requirements.txt
# # JAX GPU (wheels only available for Linux)
# --find-links https://storage.googleapis.com/jax-releases/jax_releases.html
# jax[cuda]; sys_platform == "linux"
#
# # For JAX on Windows; add the following option when installing with pip (--use-deprecated legacy-resolver)
# --find-links https://whls.blob.core.windows.net/unstable/index.html
# jax[cuda]==0.3.2; sys_platform == "win32"

numpy
einops

# Experiment Tracking
wandb
requests

# Optimization Library
optax

# Inference
# tf-nightly
onnxruntime
tf2onnx

# Config
pyyaml
